when in doubt, re-write most of it :D
