+++ htmlContent/omnisharp-codeactions-editor-template.html
