+++ htmlContent/omnisharp-findreferences-template.html
@@ -1,0 +1,7 @@
+<div class="omnisharp-references">
+<div class="inline-editor-header"><div class="filename">References</div></div>
+    <div class="references">
+        <ul>
+        </ul>
+    </div>
+</div>`
